pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mmdvm-statemachine"
version = "0.1.0"
description = "Real-time state machine and API service for MMDVMHost"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "See MMDVMHost project for license details"}
authors = [
    {name = "MMDVM State Machine Contributors"}
]
keywords = [
    "mmdvm",
    "amateur-radio",
    "ham-radio",
    "dmr",
    "dstar",
    "p25",
    "nxdn",
    "ysf",
    "api",
    "websocket",
    "state-machine"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Communications :: Ham Radio",
    "Topic :: System :: Monitoring",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: POSIX :: Linux",
    "Environment :: No Input/Output (Daemon)",
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "PyYAML>=6.0",
    "inotify-simple>=1.3.5",
    "websockets>=12.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
    "pylint>=2.17.0",
    "types-PyYAML>=6.0.0",
]

performance = [
    "uvloop>=0.19.0; sys_platform != 'win32'",
]

all = [
    "mmdvm-statemachine[dev,performance]",
]

[project.urls]
Homepage = "https://github.com/yourusername/mmdvm-statemachine"
Documentation = "https://github.com/yourusername/mmdvm-statemachine/blob/main/README.md"
Repository = "https://github.com/yourusername/mmdvm-statemachine"
Issues = "https://github.com/yourusername/mmdvm-statemachine/issues"

[project.scripts]
mmdvm-statemachine = "mmdvm_statemachine.__main__:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["mmdvm_statemachine*"]
exclude = ["tests*"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--showlocals",
]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
strict_optional = true

[[tool.mypy.overrides]]
module = [
    "inotify_simple.*",
    "uvloop.*",
]
ignore_missing_imports = true

[tool.pylint.messages_control]
max-line-length = 100
disable = [
    "C0111",  # missing-docstring
    "C0103",  # invalid-name
    "R0903",  # too-few-public-methods
    "R0913",  # too-many-arguments
]

[tool.pylint.design]
max-args = 8
max-attributes = 12
